+++ src/booklab/booklabd/__init__.py
@@ -1,12 +1,13 @@
 
 """ booklabd - Booklab API server module
-Main functionalities:
-- operate JSON database file
-- provide necessary http routes
 
-Architecture; HTTP REST
+Main functionalities:
+* operate & manage JSON database file
+* provide necessary http API routes fir book functionalities that need write & dynamic (run-time) behavior
+
+Architecture; HTTP WSGI
+
 Author: Petre Iordanescu (petre.iordanescu@gmail.com)
-Created; 21-Aug-2025
 """
 
 # general imports
@@ -14,7 +15,7 @@ import os
 import pysondb
 # booklab package imports
 from booklab.booklabd.app_init import init_app
-from booklab.booklabd.db_init import init_db
+from booklab.booklib.db_init import init_db
 from booklab import PROJECT_ROOT
 from booklab import PACKAGE_ROOT
 from booklab import CONF_ROOT