+++ src/booklab/booklib/db_init.py
@@ -1,8 +1,9 @@
-"""**db_init** module that initialize `booklabd` database (more JSON files)
+"""
+Module that initialize `booklabd` database (more JSON files)
+
 This module is designed to initialize database objects `db_books` and `db_system`.
 
 Author: Petre Iordanescu (petre.iordanescu@gmail.com)
-Created: Sep.2025
 """
 
 import pysondb