+++ src/booklab/conf/__init__.py
@@ -1,10 +1,15 @@
-"""**conf** - configurations module
-Various booklab system components configuration files.
-- _Note 1:_ files wirh `.py` extension are considered Python files and can be imported accordingly as longas they are Python-code correct.
-- _Note 2:_ all files from here are assumed to be _text files_. Any deviation from this rule is assumed to be treated accordingly in code.
+"""
+Cnfigurations component contains various system components configuration files:
+
+- `gunicorn` WSGI web server
+- `nginx` local proxy HTTP
+
+Notes / remarks:
+
+- files wirh `.py` extension are considered Python files and can be imported accordingly as longas they are valid Python-code.
+- all files from here are assumed to be _text files_. Any deviation from this rule is assumed to be treated accordingly in code.
 
 Author; Petre Iordanescu (petre.iordanescu@gmail.com)
-Created; Aug.2025
 """