+++ CHANGELOG.md
@@ -7,11 +7,37 @@
 
 ## ... #TODO [unreleased] BCAT extend book command center (...wip...)
 * tbd... #TODO-ASAP__focus on Jupiter__
-* tbd... init TMPL with a start `index.md` file && maybe 1, 2 examples markdown as "starter samples"
 * tbd... ref command "Verificare" (code-name `bstatus`):
-    * [ ] VIEW CONFIG FILE (to display `book_mkdocs.yml`)
-    * [ ] SHOW STATUS of assembly (what did in `0.2.4`), check if exists `book_mkdocs.yml.tmpl` that meaning `bkcmd_assembly...sh` was xecuted but not `bkcmd_render...py`
-    * [ ] update 810-DSGN with `bstatus` component
+
+
+
+
+
+## [0.5] - 22.sep.2025
+Implement `bstatus` functionality
+* [0.5rc1] make menu link in "technical area" to data/book_catalog.json (see 0.5a0)
+* [0.5rc0] 250922
+    *  update all redirects using EXT_PATH
+    * created gh-workflow to publish booklab package on PyPi
+    * bstatus_template.md increase "clarity" of cmd  buttons
+* [0.5b3] 250921
+    * update bstatus_template to a presentable firm of information
+    * import external URLs at booklab.__init__.py level
+    * construct a FULL_EXT_URL from parts
+    * create booklab-config-file  `.../booklab/conf/booklab_ext_url.py` for external access URL parts
+* [0.5b2] 250919 render bstatus_tenplate.md as a first proof of concept and test
+* [0.5b1] create lib function `getBook(bk_code)`
+* [0.5b0] `booklab` refactor: move db_init.py from module `booklabd` to module `booklib`
+* [0.5a2] 250917 create `booklab.booklib` module for Booklab app general purpose functions
+* [0.5a1] 250917
+    * bcat/bcat_template.md -copy-to-> bstatus_bstat_template.md
+    * update page header and explana in text
+* [0.5a0] 250917
+    * clean all technical documentation and modules docstring
+    * update 810-DSGN with `bstatus` component and change text CGI with WSGI
+* [0.5.dev2] 250916 update `books_catalog.json` with data requured by `bstatus` functionality
+* [0.5.dev1] 250916 clean prj docstrings and dirs & files
+* [0.5.dev0] 250916 init 0.5 version